+++ message_segment.h
@@ -26,11 +26,11 @@ public:
         fin = 0x8,
     };
 
-    message_segment(uint8_t source_port, uint8_t destination_port, uint16_t sequence_num, uint16_t ack_num, uint8_t type, uint8_t flags, const std::vector<uint8_t> &message);
+    message_segment(uint16_t source_port, uint16_t destination_port, uint16_t sequence_num, uint16_t ack_num, uint8_t type, uint8_t flags, const std::vector<uint8_t> &message);
     message_segment(const std::vector<uint8_t> &segment);
 
-    uint8_t get_source_port() const;
-    uint8_t get_destination_port() const;
+    uint16_t get_source_port() const;
+    uint16_t get_destination_port() const;
     uint16_t get_sequence_num() const;
     uint16_t get_ack_num() const;
     uint8_t get_message_type() const;
@@ -45,12 +45,12 @@ public:
     operator std::vector<uint8_t>() const;
 
 private:
-    uint8_t source_port;
-    uint8_t destination_port;
+    uint16_t source_port;
+    uint16_t destination_port;
     uint16_t sequence_num;
     uint16_t ack_num;
     uint8_t flags;  // bits 0-3: message flags, bits 4-7: message type
-    std::vector<uint8_t> message;  // TODO: limit to 93 bytes (i.e. max_rf_data_size - message_segment overhead)
+    std::vector<uint8_t> message;  // TODO: limit to 91 bytes (i.e. max_rf_data_size - message_segment overhead)
 };
 
 #endif
